Fixed duplicate issue with the Python 2 script.
By combining the image ID with the filename I've essentially made it so that there are fewer duplicate images in the resulting output and there's no need to implement randomization with the filenames because files with duplicate filenames will have different image IDs which solves that issue.

Before this the duplicates were simply mixed in with some randomly generated characters to avoid overwriting files with similar filenames but different contents which resulted in more duplicates (even with my attempt at clearing out duplicate entries using sets in Python). This hasn't been tested yet and I'll probably come back with more commits after testing this with the latest version of Python 2.7.