Bugfix Attempt #1
Fixed the text gathering function so that it can be imported into a database since CSVs are well known and supported.

This is labelled as an attempt since there were some development issues along the way with this release; so don't hold back on the issues because I can't test all possible combinations of systems and servers by myself.

I also added some documentation again to the file and I improved the separation of the web scraping classes between the Python 2.7 and Python 3 interpreter support.